requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
requests==2.28.2
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.14
yarl==1.8.2
